backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.4
python-decouple==3.8
psycopg2-binary==2.9.9
django-extensions==3.2.3
websockets==12.0
daphne==4.0.0